+++ example/a.ts
@@ -1,37 +1,17 @@
-import { Elysia, t } from '../src'
-import { retrieveRootParamters, sucrose } from '../src/sucrose'
+import { Elysia } from '../src'
+import { req } from '../test/utils'
 
-// const parameter = '({ hello: { a }, path })'
-// const result = retrieveRootParamters(parameter)
-
-// console.log(result)
-
-console.log(
-	sucrose({
-		handler: function ({ query }) {
-			query.a
-		},
-		afterHandle: [],
-		beforeHandle: [],
-		error: [
-			function a({
-				query,
-				query: { a, c: d },
-				headers: { hello },
-				...rest
-			}) {
-				query.b
-				rest.query.e
-			},
-			({ query: { f } }) => {}
-		],
-		mapResponse: [],
-		onResponse: [],
-		parse: [],
-		request: [],
-		start: [],
-		stop: [],
-		trace: [],
-		transform: []
+const app = new Elysia()
+	.get('/', function* ({ set }) {
+		return 'hello'
 	})
-)
+	.get('/json', function* ({ set }) {
+		return { hello: 'world' }
+	})
+
+const response = await Promise.all([
+	app.handle(req('/')),
+	app.handle(req('/json'))
+])
+
+console.log(response[0].headers.get('content-type'))