+++ src/types.ts
@@ -821,9 +821,16 @@ export type ErrorHandler<
 						code: 'UNKNOWN'
 						error: Readonly<Error>
 						set: Context['set']
-						a: Singleton
-					} & Partial<Singleton['derive'] & Ephemeral['derive'] & Volatile['derive']> &
-						Partial<Singleton['derive'] & Ephemeral['resolve'] & Volatile['resolve']>
+					} & Partial<
+						Singleton['derive'] &
+							Ephemeral['derive'] &
+							Volatile['derive']
+					> &
+						Partial<
+							Singleton['derive'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						>
 			  >
 			| Prettify<
 					{
@@ -831,8 +838,14 @@ export type ErrorHandler<
 						code: 'VALIDATION'
 						error: Readonly<ValidationError>
 						set: Context['set']
-					} & NeverKey<Singleton['derive'] & Ephemeral['derive'] & Volatile['derive']> &
-						NeverKey<Singleton['derive'] & Ephemeral['resolve'] & Volatile['resolve']>
+					} & Singleton['derive'] &
+						Ephemeral['derive'] &
+						Volatile['derive'] &
+						NeverKey<
+							Singleton['derive'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						>
 			  >
 			| Prettify<
 					{
@@ -840,8 +853,16 @@ export type ErrorHandler<
 						code: 'NOT_FOUND'
 						error: Readonly<NotFoundError>
 						set: Context['set']
-					} & NeverKey<Singleton['derive'] & Ephemeral['derive'] & Volatile['derive']> &
-						NeverKey<Singleton['derive'] & Ephemeral['resolve'] & Volatile['resolve']>
+					} & NeverKey<
+						Singleton['derive'] &
+							Ephemeral['derive'] &
+							Volatile['derive']
+					> &
+						NeverKey<
+							Singleton['derive'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						>
 			  >
 			| Prettify<
 					{
@@ -849,8 +870,14 @@ export type ErrorHandler<
 						code: 'PARSE'
 						error: Readonly<ParseError>
 						set: Context['set']
-					} & NeverKey<Singleton['derive'] & Ephemeral['derive'] & Volatile['derive']> &
-						NeverKey<Singleton['derive'] & Ephemeral['resolve'] & Volatile['resolve']>
+					} & Singleton['derive'] &
+						Ephemeral['derive'] &
+						Volatile['derive'] &
+						NeverKey<
+							Singleton['derive'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						>
 			  >
 			| Prettify<
 					{
@@ -858,8 +885,16 @@ export type ErrorHandler<
 						code: 'INTERNAL_SERVER_ERROR'
 						error: Readonly<InternalServerError>
 						set: Context['set']
-					} & Partial<Singleton['derive'] & Ephemeral['derive'] & Volatile['derive']> &
-						Partial<Singleton['derive'] & Ephemeral['resolve'] & Volatile['resolve']>
+					} & Partial<
+						Singleton['derive'] &
+							Ephemeral['derive'] &
+							Volatile['derive']
+					> &
+						Partial<
+							Singleton['derive'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						>
 			  >
 			| Prettify<
 					{
@@ -867,8 +902,16 @@ export type ErrorHandler<
 						code: 'INVALID_COOKIE_SIGNATURE'
 						error: Readonly<InvalidCookieSignature>
 						set: Context['set']
-					} & NeverKey<Singleton['derive'] & Ephemeral['derive'] & Volatile['derive']> &
-						NeverKey<Singleton['derive'] & Ephemeral['resolve'] & Volatile['resolve']>
+					} & NeverKey<
+						Singleton['derive'] &
+							Ephemeral['derive'] &
+							Volatile['derive']
+					> &
+						NeverKey<
+							Singleton['derive'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						>
 			  >
 			| Prettify<
 					{
@@ -879,8 +922,16 @@ export type ErrorHandler<
 							set: Context['set']
 						}
 					}[keyof T] &
-						Partial<Singleton['derive'] & Ephemeral['derive'] & Volatile['derive']> &
-						Partial<Singleton['derive'] & Ephemeral['resolve'] & Volatile['resolve']>
+						Partial<
+							Singleton['derive'] &
+								Ephemeral['derive'] &
+								Volatile['derive']
+						> &
+						Partial<
+							Singleton['derive'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						>
 			  >
 		)
 ) => any | Promise<any>