+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.1.3 - 17 Jul 2024
+Change:
+- sucrose: exact inference name
+
+Bug fix:
+- `onError` with scope not being able to infer context type
+
 # 1.1.2 - 16 Jul 2024
 Bug fix:
 - [#724](https://github.com/elysiajs/elysia/issues/724), [bun#12594](https://github.com/oven-sh/bun/issues/12594) sucrose: possibly fix `bun build --compile` not being able to infer first, and last context parameter